10. RAG-WEB/app/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10
openai>=1.1.0
qdrant-client>=1.6.0
beautifulsoup4>=4.9.3
requests>=2.31.0
pydantic>=2.0.0
python-multipart>=0.0.6 